+++ src/main/java/sbitneva/controller/EditController.java
@@ -1,0 +1,4 @@
+package sbitneva.controller;
+
+public class EditController {
+}
